More playing w/ pygame + simpy
Pretty fun stuff. Got some super-basic graphics tied into the simpy simulation.

Started some abstractions to make it easier to write drawable components; it's a little tricky to figure out when to redraw.

A blanket "draw after yield" doesn't work because we might be changing state and then waiting.

Drawing with the workflow feels a little bad because it's mixing drawing and simulation concerns, but it feels like a LOT more work to split it up.